ImageBinarization.jl Documentation

ImageBinarization.jl Documentation

ImageBinarization.binarize — Method.

imgb = binarize(Otsu(), img)
binarize!(Otsu(), img)

Under the assumption that the histogram is bimodal the threshold is set so that the resultant inter-class variance is maximal.

Arguments

The function arguments are described in more detail below.

img

An AbstractGray image which is to be binarized into background (Gray(0)) and foreground (Gray(1)) pixels.

Reference

Nobuyuki Otsu (1979). "A threshold selection method from gray-level histograms". IEEE Trans. Sys., Man., Cyber. 9 (1): 62–66. doi:10.1109/TSMC.1979.4310076

source

ImageBinarization.binarize — Method.

imgb = binarize(Polysegment(), img)
binarize!(Polysegment(), img)

Uses the polynomial segmentation technique to group the image pixels into two categories (foreground and background).

Details

The approach involves constructing a univariate second-degree polynomial such that the two roots of the polynomial represent the graylevels of two cluster centers (i.e the foreground and background). Pixels are then assigned to the foreground or background depending on which cluster center is closest.

Arguments

The function arguments are described in more detail below.

img

An AbstractGray image which is to be binarized into background (Gray(0)) and foreground (Gray(1)) pixels.

Reference

[1] R. E. Vidal, "Generalized Principal Component Analysis (GPCA): An Algebraic Geometric Approach to Subspace Clustering and Motion Segmentation." Order No. 3121739, University of California, Berkeley, Ann Arbor, 2003.

source

ImageBinarization.binarize! — Method.

t = binarize!(Otsu(), img)

Same as binarize except that it modifies the image that was passed as an argument.

source

ImageBinarization.binarize! — Method.

t = binarize!(Polysegment(), img)

Same as binarize except that it modifies the image that was passed as an argument.

source