Function References

Function References

General function

Algorithms

Adaptive Threshold

Niblack

Polysegment

Sauvola

Algorithms that utilizes single histogram-threshold

General function

ImageBinarization.BinarizationAPI.binarize — Function

binarize([T::Type,] img, f::AbstractImageBinarizationAlgorithm, args...; kwargs...)

Binarize img using algorithm f.

Output

The return image img₀₁ is an Array{T}.

If T is not specified, then it's inferred as Gray{eltype(eltype(img))}, which is Gray{N0f8} for img of type Array{N0f8} and Array{Gray{N0f8}}, and Gray{Float32} for img of type Array{Float32} and Array{Gray{Float32}}

Examples

Just simply pass the input image and algorithm to binarize

img₀₁ = binarize(img, f)

This reads as "binarize image img using binarization algorithm f".

You can also explicitly specify the return type:

img₀₁_float32 = binarize(Gray{Float32}, img, f)

See also binarize! for in-place binarization.

ImageBinarization.BinarizationAPI.binarize! — Function

binarize!([out,] img, f::AbstractImageBinarizationAlgorithm, args...; kwargs...)

Binarize img using algorithm f.

Output

If out is specified, it will be changed in place. Otherwise img will be changed in place.

Examples

Just simply pass an algorithm to binarize!:

img₀₁ = similar(img)
binarize!(img₀₁, img, f)

For cases you just want to change img in place, you don't necessarily need to manually allocate img₀₁; just use the convenient method:

binarize!(img, f)

See also: binarize

Algorithms

ImageBinarization.BinarizationAPI.AbstractImageBinarizationAlgorithm — Type

AbstractImageBinarizationAlgorithm <: AbstractImageFilter

The root type for ImageBinarization package.

Any concrete binarization algorithm shall subtype it to support binarize and binarize! APIs.

Examples

All algorithms in ImageBinarization are called in the following pattern:

# first generate an algorithm instance
f = Otsu()

# then pass the algorithm to `binarize`
img₀₁ = binarize(img, f) # `eltype(img₀₁)` is `Gray{N0f8}`

# or use in-place version `binarize!`
img₀₁ = similar(img)
binarize!(img₀₁, img, f)

Some algorithms also receive additional information of image as an argument to infer the "best" parameters, e.g., window_size of AdaptiveThreshold.

# you could explicit specify the it
f = AdaptiveThreshold(window_size = 32)

# or infer the "best" default value from given image
img = testimage("cameraman")
f = AdaptiveThreshold(img)

For more examples, please check binarize and binarize! and concret algorithms.

Adaptive Threshold

ImageBinarization.AdaptiveThreshold — Type

AdaptiveThreshold <: AbstractImageBinarizationAlgorithm
AdaptiveThreshold([img]; [window_size,] percentage = 15)

binarize([T,] img, f::AdaptiveThreshold)
binarize!([out,] img, f::AdaptiveThreshold)

Binarize img using a threshold that varies according to background illumination.

Output

Return the binarized image as an Array{Gray{T}} of size size(img). If T is not specified, it is inferred from out and img.

Details

If the value of a pixel is t percent less than the average of an $s \times s$ window of pixels centered around the pixel, then the pixel is set to black, otherwise it is set to white.

A computationally efficient method for computing the average of an $s \times s$ neighbourhood is achieved by using an integral image integral_image.

This algorithm works particularly well on images that have distinct contrast between background and foreground. See [1] for more details.

Arguments

The function argument is described in more detail below.

img::AbstractArray

The image that need to be binarized. The image is automatically converted to Gray in order to construct the requisite graylevel histogram.

You can also pass img to AdaptiveThreshold to automatically infer the "best" window_size.

Options

Various options for the parameters of AdaptiveThreshold, binarize and binarize! are described in more detail below.

Choices for percentage

You can specify an integer for the percentage (denoted by t in [1]) which must be between 0 and 100. Default: 15

Choices for window_size

The argument window_size (denoted by s in [1]) specifies the size of pixel's square neighbourhood which must be greater than zero.

If img is passed to AdaptiveThreshold constructor, then window_size is infered as the integer closest to 1/8 of the average of the width and height of img.

Examples

using TestImages

img = testimage("cameraman")
f = AdaptiveThreshold(window_size = 16)
img₀₁ = binarize(img, f)

f = AdaptiveThreshold(img) # infer the best `window_size` using `img`
img₀₁ = binarize(img, f)

See also binarize! for in-place operation.

References

[1] Bradley, D. (2007). Adaptive Thresholding using Integral Image. Journal of Graphic Tools, 12(2), pp.13-21. doi:10.1080/2151237x.2007.10129236

Niblack

ImageBinarization.Niblack — Type

Niblack <: AbstractImageBinarizationAlgorithm
Niblack(; window_size = 7, bias = 0.2)

binarize([T,] img, f::Niblack)
binarize!([out,] img, f::Niblack)

Applies Niblack adaptive thresholding [1] under the assumption that the input image is textual.

Output

Return the binarized image as an Array{Gray{T}} of size size(img). If T is not specified, it is inferred from out and img.

Details

The input image is binarized by varying the threshold across the image, using a modified version of Niblack's algorithm [2]. A threshold $T$ is defined for each pixel based on the mean $m$ and standard deviation $s$ of the intensities of neighboring pixels in a window around it. This threshold is given by

\[T(x,y) = m(x,y) + k \cdot s(x,y),\]

where $k$ is a user-defined parameter weighting the influence of the standard deviation on the value of $T$.

Note that Niblack's algorithm is highly sensitive to variations in the gray values of background pixels, which often exceed local thresholds and appear as artifacts in the binarized image. The Sauvola algorithm included in this package implements an attempt to address this issue [2].

Arguments

img

An image which is binarized according to a per-pixel adaptive threshold into background (0) and foreground (1) pixel values.

bias::Real (denoted by $k$ in the publication)

A user-defined biasing parameter on threshold. This can take negative values. Larger bias encourages more black pixels in the output.

window_size::Integer (denoted by $w$ in the publication)

The threshold for each pixel is a function of the distribution of the intensities of all neighboring pixels in a square window around it. The side length of this window is $2w + 1$, with the target pixel in the center position.

If not specified, window_size is 7.

Example

Binarize the "cameraman" image in the TestImages package.

using TestImages, ImageBinarization

img = testimage("cameraman")
img₀₁ = binarize(img, Niblack())

References

[1] Wayne Niblack (1986). An Introduction to Image Processing. Prentice-Hall, Englewood Cliffs, NJ: 115-16. [2] J. Sauvola and M. Pietikäinen (2000). "Adaptive document image binarization". Pattern Recognition 33 (2): 225-236. doi:10.1016/S0031-3203(99)00055-2

Polysegment

ImageBinarization.Polysegment — Type

Polysegment <: AbstractImageBinarizationAlgorithm
Polysegment()

binarize([T,] img, f::Polysegment)
binarize!([out,] img, f::Polysegment)

Uses the polynomial segmentation technique to group the image pixels into two categories (foreground and background).

Output

Return the binarized image as an Array{Gray{T}} of size size(img). If T is not specified, it is inferred from out and img.

Details

The approach involves constructing a univariate second-degree polynomial such that the two roots of the polynomial represent the graylevels of two cluster centers (i.e the foreground and background). Pixels are then assigned to the foreground or background depending on which cluster center is closest.

Arguments

The function argument is described in more detail below.

img::AbstractArray

The image that needs to be binarized. The image is automatically converted to Gray.

Example

Binarize the "cameraman" image in the TestImages package.

using TestImages, ImageBinarization

img = testimage("cameraman")
img_binary = binarize(img, Polysegment())

Reference

R. E. Vidal, "Generalized Principal Component Analysis (GPCA): An Algebraic Geometric Approach to Subspace Clustering and Motion Segmentation." Order No. 3121739, University of California, Berkeley, Ann Arbor, 2003.

Sauvola

ImageBinarization.Sauvola — Type

Sauvola <: AbstractImageBinarizationAlgorithm
Sauvola(; bias = 0.2, window_size=7)

binarize([T,] img, f::Sauvola)
binarize!([out,] img, f::Sauvola)

Applies Sauvola–Pietikäinen adaptive image binarization [1] under the assumption that the input image is textual.

Output

Return the binarized image as an Array{Gray{T}} of size size(img). If T is not specified, it is inferred from out and img.

Details

The input image is binarized by varying the threshold across the image, using a modified version of Niblack's algorithm [2]. Niblack's approach was to define a threshold $T$ for each pixel based on the mean $m$ and standard deviation $s$ of the intensities of neighboring pixels in a window around it, given by

\[T(x,y) = m(x,y) + k \cdot s(x,y),\]

where $k$ is a user-defined parameter weighting the influence of the standard deviation on the value of $T$.

Niblack's algorithm is highly sensitive to variations in the gray values of background pixels, which often exceed local thresholds and appear as artifacts in the binarized image. Sauvola and Pietikäinen [1] introduce the dynamic range $R$ of the standard deviation (i.e. its maximum possible value in the color space), such that the threshold is given by

\[T(x,y) = m(x,y) \cdot \left[ 1 + k \cdot \left( \frac{s(x,y)}{R} - 1 \right) \right]\]

This adaptively amplifies the contribution made by the standard deviation to the value of $T$.

The Sauvola–Pietikäinen algorithm is implemented here using an optimization proposed by Shafait, Keysers and Breuel [3], in which integral images are used to calculate the values of $m$ and $s$ for each pixel in constant time. Since each of these data structures can be computed in a single pass over the source image, runtime is significantly improved.

Arguments

img

An image which is binarized according to a per-pixel adaptive threshold into background (0) and foreground (1) pixel values.

window_size::Integer (denoted by $w$ in the publication)

The threshold for each pixel is a function of the distribution of the intensities of all neighboring pixels in a square window around it. The side length of this window is $2w + 1$, with the target pixel in the center position.

bias::Real (denoted by $k$ in the publication)

A user-defined biasing parameter. This can take negative values, though values in the range [0.2, 0.5] are typical. According to [1], this algorithm is not too sensitive to the value of $k$`.

Example

Binarize the "cameraman" image in the TestImages package.

using TestImages, ImageBinarization

img = testimage("cameraman")
img_binary = binarize(img, Sauvola(window_size = 9, bias = 0.2))

References

J. Sauvola and M. Pietikäinen (2000). "Adaptive document image binarization". Pattern Recognition 33 (2): 225-236. doi:10.1016/S0031-3203(99)00055-2

Wayne Niblack (1986). An Introduction to Image Processing. Prentice-Hall, Englewood Cliffs, NJ: 115-16.

Faisal Shafait, Daniel Keysers and Thomas M. Breuel (2008). "Efficient implementation of local adaptive thresholding techniques using integral images". Proc. SPIE 6815, Document Recognition and Retrieval XV, 681510 (28 January 2008). doi:10.1117/12.767755

Algorithms that utilizes single histogram-threshold

The core functionality of these algorithms are supported by HistogramThresholding.jl

ImageBinarization.SingleHistogramThreshold — Type

SingleHistogramThreshold <: AbstractImageBinarizationAlgorithm
SingleHistogramThreshold(alg::ThresholdAlgorithm; nbins=256)

binarize([T,] img, f::ThresholdAlgorithm; nbins=256)
binarize!([out,] img, f::ThresholdAlgorithm; nbins=256)

Binarizes the image img using the threshold found by given threshold finding algorithm alg.

Output

Return the binarized image as an Array{Gray{T}} of size size(img). If T is not specified, it is inferred from out and img.

Arguments

The function argument is described in more detail below.

img::AbstractArray

The image that needs to be binarized. The image is automatically converted to Gray in order to construct the requisite graylevel histogram.

alg::ThresholdAlgorithm

ThresholdAlgorithm is an Abstract type defined in ThresholdAlgorithm.jl, it provides various threshold finding algorithms:

HistogramThresholding.Balanced

HistogramThresholding.Entropy

HistogramThresholding.Intermodes

HistogramThresholding.MinimumError

HistogramThresholding.MinimumIntermodes

HistogramThresholding.Moments

HistogramThresholding.Otsu

HistogramThresholding.UnimodalRosin

HistogramThresholding.Yen

For the more detailed explaination and the construction, please refer to each concrete algorithm. For example, type ?Otsu in REPL will give you more details on how to use Otsu methods.

nbins::Integer

The number of discrete bins that used to build the histogram. A smaller nbins could possibly gives a less noisy, or in other words, a smoother output. The default value is 256.

Examples

All the usage follows the same pattern, take Otsu as an example:

using TestImages, ImageBinarization

img = testimage("cameraman")
img_binary = binarize(img, Otsu())

It is less convenient, but still, you could also construct a SingleHistogramThreshold by yourself:

using TestImages, ImageBinarization

img = testimage("cameraman")
f = SingleHistogramThreshold(Otsu(), nbins=256)
img_binary = binarize(img, f)